Move away from react
Made the choice to try and build this without react - there are only three pieces of moving data, and the functions that do the calculating don't rely on react at all.

Put together a small service that uses gulp to browserify/babel/package all of the javascript (so that modules can still be used) into one file, and runs a development server with livereload so that functionality isn't lost.